requirements.txt
numpy>=1.23,<1.27
opencv-python>=4.9.0.80,<4.10
pygame>=2.5
PyYAML>=6.0
